Changelog
Hi Team,

I've made a new branch called listdirbranch in the repo and made the
following changes:

1. Implemented command line interface that outputs "ftp>>" and listens
for command from user, taking appropriate actions.

2. Implemented user log in to FileZilla server. I've preserved the
'test' login on my machine. I also tested it with anonymous login on
Edinburgh server.

3. Implemented list files and directories functionality. Directories are
marked with ~ and files are marked with *

4. Implemented 'help' functionality as a part of documentation process.
When user types 'help' on the command line, he will be shown a list of
available commands. This is still under construction and should be
updated as the project proceeds.

5. Implemented log out and exit functionality. This was not a task, but
it was necessary to log out of the server before exiting the code. It
can and should be separated later and documentation should be updated
accordingly.

I'm waiting for test cases from Abhishek so that I can execute them, as
per our planned tasks, for the first story. It will be good if we can
maintain all test cases in a shared Excel on drive along with the
results. I will add the test cases for the second story when the file is
shared.

An issue that remains to be resolved is handling server response to both
login and list files. I'll work on both tomorrow.

Another feature we need to implement is password masking, where a series
of * will be displayed to the user while typing the password, instead of
the plain text that we currently have.

Let's target finishing the first two stories completely by Tuesday, at
the very least.

Please test this code whenever you guys have time and give me feedback
:)

Deven